+++ scripts/const.js
@@ -75,7 +75,6 @@ export const TEMPLATES = {
     PANEL_CHARACTER_SUMMARY: `modules/${MODULE.ID}/templates/panel-character-summary.hbs`,
     PANEL_PARTY: `modules/${MODULE.ID}/templates/panel-party.hbs`,
     PANEL_PARTY_STATS: `modules/${MODULE.ID}/templates/panel-party-stats.hbs`,
-    PANEL_NOTES: `modules/${MODULE.ID}/templates/panel-notes.hbs`,
     HANDLE_PLAYER: `modules/${MODULE.ID}/templates/handle-player.hbs`,
     HANDLE_PARTY: `modules/${MODULE.ID}/templates/handle-party.hbs`,
     CHAT_CARD: `modules/${MODULE.ID}/templates/chat-cards.hbs`,